T-012: Convert HandDashboard and PlayerGrid to TSX (aia-core-13lt)
