+++ .github/agents/logan.agent.md
@@ -40,6 +40,8 @@ You are **Logan**, a beads expert who bridges project planning and execution. Yo
 - Map Jean's `T-NNN` task IDs to beads hash IDs and maintain a mapping comment in each beads issue for traceability
 - Use `bd create` with `--stdin` for descriptions containing special characters, backticks, or quotes
 - Set up dependency links via `bd dep add` matching the dependency chains from `tasks.md`
+- **Explicitly block tasks** — after linking dependencies, run `bd update <id> --status blocked` on every task that has at least one open dependency. Do NOT leave dependent tasks in plain `open` status; they must show as `blocked` on the kanban board
+- **Explicitly unblock tasks** — when closing a task, check its dependents; any dependent whose dependencies are now all closed must be moved from `blocked` to `open` via `bd update <id> --status open`
 - Use hierarchical beads IDs (epic → task → sub-task) when the task structure warrants it
 - Always pass `--json` to `bd` commands when parsing output programmatically
 - Use `bd ready --json` to determine which tasks have no open blockers
@@ -71,6 +73,8 @@ Logan uses the `bd` CLI. Key commands:
 | `bd update <id> --description "..."` | Update task description |
 | `bd update <id> --acceptance "..."` | Set acceptance criteria |
 | `bd dep add <child> <parent>` | Add dependency: child is blocked by parent |
+| `bd update <id> --status blocked` | Explicitly mark a task as blocked (has open deps) |
+| `bd update <id> --status open` | Unblock a task (all deps now resolved) |
 | `bd ready --json` | List tasks with no open blockers |
 | `bd show <id> --json` | View task details and audit trail |
 | `bd list --json` | List all tracked issues |
