+++ docs/agent/reviews/cycle-3-aia-core-vbsk-2026-04-11.md
@@ -1,0 +1,45 @@
+# Code Review Report — Cycle 3: aia-core-vbsk (T-003)
+
+**Reviewer:** Scott
+**Date:** 2026-04-11
+**Epic:** aia-core-dthg
+**Task:** aia-core-vbsk (T-003: Configure tsconfig.json — strict)
+**Cycle:** 3
+
+---
+
+## Acceptance Criteria
+
+| AC | Status | Evidence |
+|---|---|---|
+| 1. `"strict": true` | MET | `frontend/tsconfig.json` line 19 |
+| 2. `"jsx": "react-jsx"` | MET | `frontend/tsconfig.json` line 16 |
+| 3. `"moduleResolution": "bundler"` | MET | `frontend/tsconfig.json` line 10 |
+| 4. `"include": ["src"]` / `"exclude": ["node_modules", "dist"]` | MET | `frontend/tsconfig.json` lines 26–27 |
+| 5. `npx tsc --noEmit` runs | MET | Commit message confirms TS18003 (expected with no `.ts` files) |
+
+---
+
+## Findings
+
+### MEDIUM
+
+**Preact/React toolchain mismatch (pre-existing)**
+- **File:** `frontend/vite.config.js` line 2
+- **Description:** `vite.config.js` imports `@preact/preset-vite`, but `tsconfig.json` configures `"jsx": "react-jsx"` and `package.json` lists `react`, `react-dom`, and `@vitejs/plugin-react`. When `.tsx` files are introduced, Vite will transform JSX for Preact while TypeScript expects React types.
+- **Suggestion:** Update `vite.config.js` to use `@vitejs/plugin-react` instead of `@preact/preset-vite`. This is outside the scope of T-003 but should be tracked.
+
+### LOW
+
+**Missing `forceConsistentCasingInFileNames`**
+- **File:** `frontend/tsconfig.json`
+- **Description:** The official Vite React-TS template includes `"forceConsistentCasingInFileNames": true` for cross-platform safety. Not strictly required on Linux.
+- **Suggestion:** Add `"forceConsistentCasingInFileNames": true` to compilerOptions for portability.
+
+---
+
+## Summary
+
+All 5 acceptance criteria are met. The `tsconfig.json` follows Vite + React + TypeScript best practices: `isolatedModules`, `noEmit`, `moduleDetection: "force"`, bundler module resolution, and strict mode. Zero CRITICAL or HIGH findings. One MEDIUM finding is a pre-existing toolchain inconsistency outside this task's scope.
+
+**Verdict:** PASS — no critical issues.
