+++ docs/agent/reviews/cycle-6-aia-core-vcjb-2026-04-11.md
@@ -1,0 +1,64 @@
+# Code Review Report
+
+| Field | Value |
+|---|---|
+| **Target** | Task aia-core-vcjb (T-005: Set up ESLint + CSS Modules) |
+| **Epic** | aia-core-dthg |
+| **Cycle** | 6 |
+| **Reviewer** | Scott |
+| **Date** | 2026-04-11 |
+| **Commit** | 2c3e42d |
+
+---
+
+## Files Reviewed
+
+- `frontend/eslint.config.js`
+- `frontend/src/types/css.d.ts`
+- `frontend/package.json`
+- `frontend/tsconfig.json`
+
+---
+
+## Acceptance Criteria Mapping
+
+| AC# | Criterion | Status | Notes |
+|---|---|---|---|
+| 1 | ESLint config extends `@typescript-eslint/recommended` and `plugin:react/recommended` | **PASS** | Uses `tsPlugin.configs['flat/recommended']` spread and `reactPlugin.configs.flat.recommended.rules` — correct flat config equivalents for ESLint 9.x |
+| 2 | `src/types/css.d.ts` declares module type for `*.module.css` | **PASS** | Correctly declares `readonly` string-keyed classes with default export |
+| 3 | `npm run lint` passes on config files | **PARTIAL** | Config files (`.ts`/`.tsx`) lint cleanly, but `npm run lint` exits non-zero due to 4 pre-existing errors in legacy `.js` files under `src/` |
+| 4 | Sample `.module.css` importable in `.tsx` without TS errors | **PARTIAL** | Type declaration is correct and within `tsconfig.json` `include` scope, but no sample import exists to prove end-to-end |
+
+---
+
+## Findings
+
+### MEDIUM
+
+**M1 — `npm run lint` exits non-zero due to legacy JS errors**
+- **Location:** `frontend/src/scenes/communityCards.js:29`, `frontend/src/scenes/holeCards.js:7,187`, `frontend/src/views/playbackView.js:95`
+- **Description:** The `lint` script targets all of `src/` and reports 4 errors in pre-existing `.js` files (`no-unused-vars`, `no-unused-expressions`). This will cause CI failures if a lint gate is added.
+- **Suggestion:** Either fix the 4 lint errors in the JS files, or scope the lint script to `src/**/*.{ts,tsx}` to exclude legacy JS until it is migrated. Example: `"lint": "eslint 'src/**/*.{ts,tsx}' --no-warn-ignored"`
+
+### LOW
+
+**L1 — No sample `.module.css` import to demonstrate AC4**
+- **Description:** AC4 asks for a sample `.module.css` file importable in `.tsx` without TS errors. The type declaration at `src/types/css.d.ts` is correctly shaped and within the `tsconfig.json` `include` scope, so it *should* work — but no sample file was created to prove it.
+- **Suggestion:** Add a minimal `src/App.module.css` with one class and import it in `App.tsx` to verify. Can be deferred if a future task will add CSS modules.
+
+**L2 — React version auto-detection log noise**
+- **Description:** `settings.react.version: 'detect'` is correct, but when `eslint-plugin-react` detects the version at lint time it emits an info log. Not a bug, just cosmetic.
+- **Suggestion:** No action needed; noted for awareness only.
+
+---
+
+## Summary
+
+| Severity | Count |
+|---|---|
+| CRITICAL | 0 |
+| HIGH | 0 |
+| MEDIUM | 1 |
+| LOW | 2 |
+
+**Verdict:** No critical or high findings. The ESLint flat config is well-structured and correctly adapted for ESLint 9.x. The one medium finding (lint exit code) should be resolved before adding a CI lint gate. Overall, the implementation satisfies the intent of all four acceptance criteria with minor gaps.
