+++ frontend/src/dealer/DealerApp.jsx
@@ -2,8 +2,8 @@ import { useReducer, useState, useEffect } from 'preact/hooks';
 import { reducer, initialState, validateOutcomeStreets } from './dealerState.ts';
 import { GameSelector } from './GameSelector.tsx';
 import { GameCreateForm } from './GameCreateForm.tsx';
-import { HandDashboard } from './HandDashboard.jsx';
-import { PlayerGrid } from './PlayerGrid.jsx';
+import { HandDashboard } from './HandDashboard.tsx';
+import { PlayerGrid } from './PlayerGrid.tsx';
 import { CameraCapture } from './CameraCapture.jsx';
 import { DetectionReview } from './DetectionReview.jsx';
 import { OutcomeButtons } from './OutcomeButtons.jsx';
