--- frontend/src/dealer/PlayerGrid.test.jsx
@@ -1,309 +1,0 @@
-/** @vitest-environment happy-dom */
-import { describe, it, expect, vi } from 'vitest';
-import { render } from 'preact';
-import { PlayerGrid } from './PlayerGrid.jsx';
-
-function renderToContainer(vnode) {
-  const container = document.createElement('div');
-  render(vnode, container);
-  return container;
-}
-
-const emptyCommunity = { flop1: null, flop2: null, flop3: null, flopRecorded: false, turn: null, turnRecorded: false, river: null, riverRecorded: false };
-
-const defaultProps = {
-  players: [
-    { name: 'Alice', recorded: false, status: 'playing', outcomeStreet: null },
-    { name: 'Bob', recorded: true, status: 'won', outcomeStreet: 'river' },
-    { name: 'Carol', recorded: false, status: 'folded', outcomeStreet: 'flop' },
-    { name: 'Dave', recorded: false, status: 'lost', outcomeStreet: 'turn' },
-    { name: 'Eve', recorded: false, status: 'not_playing', outcomeStreet: null },
-  ],
-  community: emptyCommunity,
-  onTileSelect: vi.fn(),
-  onBack: vi.fn(),
-  canFinish: false,
-  onFinishHand: vi.fn(),
-};
-
-describe('PlayerGrid', () => {
-  it('renders player name and status text on each tile', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const rows = container.querySelectorAll('[data-testid^="player-row-"]');
-    expect(rows.length).toBe(5);
-
-    expect(rows[0].textContent).toContain('Alice');
-    expect(rows[0].textContent).toContain('playing');
-
-    expect(rows[1].textContent).toContain('Bob');
-    expect(rows[1].textContent).toContain('won');
-    expect(rows[1].textContent).toContain('river');
-
-    expect(rows[2].textContent).toContain('Carol');
-    expect(rows[2].textContent).toContain('folded');
-    expect(rows[2].textContent).toContain('flop');
-
-    expect(rows[3].textContent).toContain('Dave');
-    expect(rows[3].textContent).toContain('lost');
-    expect(rows[3].textContent).toContain('turn');
-
-    expect(rows[4].textContent).toContain('Eve');
-    expect(rows[4].textContent).toContain('not playing');
-  });
-
-  it('sets background color based on player status', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const rows = container.querySelectorAll('[data-testid^="player-row-"]');
-
-    expect(rows[0].style.backgroundColor).toBe('#ffffff'); // playing = white
-    expect(rows[1].style.backgroundColor).toBe('#bbf7d0'); // won = green
-    expect(rows[2].style.backgroundColor).toBe('#fecaca'); // folded = red
-    expect(rows[3].style.backgroundColor).toBe('#fed7aa'); // lost = orange
-    expect(rows[4].style.backgroundColor).toBe('#e5e7eb'); // not_playing = gray
-  });
-
-  it('renders player rows with adequate min-height', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const rows = container.querySelectorAll('[data-testid^="player-row-"]');
-    for (const row of rows) {
-      expect(parseInt(row.style.minHeight)).toBeGreaterThanOrEqual(48);
-    }
-  });
-
-  it('shows not_playing status without outcome button allowing change', () => {
-    const onDirectOutcome = vi.fn();
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} onDirectOutcome={onDirectOutcome} />
-    );
-    expect(container.querySelector('[data-testid="outcome-btn-Eve"]')).not.toBeNull();
-  });
-
-  it('shows ✅ on Flop tile when flop is recorded', () => {
-    const recordedCommunity = { ...emptyCommunity, flopRecorded: true, flop1: 'Ah', flop2: 'Kd', flop3: '5c' };
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} community={recordedCommunity} />,
-    );
-    const flopTile = container.querySelector('button[data-testid="flop-tile"]');
-    expect(flopTile).not.toBeNull();
-    expect(flopTile.textContent).toContain('Flop');
-    expect(flopTile.textContent).toContain('✅');
-  });
-
-  it('does not show ✅ on Flop tile when flop is not recorded', () => {
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} community={emptyCommunity} />,
-    );
-    const flopTile = container.querySelector('button[data-testid="flop-tile"]');
-    expect(flopTile.textContent).toContain('Flop');
-    expect(flopTile.textContent).not.toContain('✅');
-  });
-
-  it('shows ✅ on player tile when recorded', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const bobRow = container.querySelector('[data-testid="player-row-Bob"]');
-    expect(bobRow.textContent).toContain('✅');
-  });
-
-  it('renders a back button that calls onBack', () => {
-    const onBack = vi.fn();
-    const container = renderToContainer(<PlayerGrid {...defaultProps} onBack={onBack} />);
-    const backBtn = container.querySelector('[data-testid="back-btn"]');
-    expect(backBtn).not.toBeNull();
-    backBtn.click();
-    expect(onBack).toHaveBeenCalledTimes(1);
-  });
-
-  it('renders street tiles above the player rows', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const streetButtons = container.querySelector('[data-testid="street-buttons"]');
-    const playerList = container.querySelector('[data-testid="player-list"]');
-    expect(streetButtons).not.toBeNull();
-    expect(playerList).not.toBeNull();
-    const flopTile = container.querySelector('[data-testid="flop-tile"]');
-    expect(flopTile).not.toBeNull();
-    expect(playerList.contains(flopTile)).toBe(false);
-  });
-
-  it('renders each player as a row with name column and status column', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const rows = container.querySelectorAll('[data-testid^="player-row-"]');
-    expect(rows.length).toBe(5);
-
-    // Each row should have the player name and status in separate sections
-    const aliceRow = container.querySelector('[data-testid="player-row-Alice"]');
-    expect(aliceRow).not.toBeNull();
-    expect(aliceRow.textContent).toContain('Alice');
-    expect(aliceRow.textContent).toContain('playing');
-  });
-
-  it('shows outcome button for players with non-playing status', () => {
-    const onDirectOutcome = vi.fn();
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} onDirectOutcome={onDirectOutcome} />
-    );
-    // Bob has status=won, Carol=folded, Dave=lost — all should show outcome button
-    expect(container.querySelector('[data-testid="outcome-btn-Bob"]')).not.toBeNull();
-    expect(container.querySelector('[data-testid="outcome-btn-Carol"]')).not.toBeNull();
-    expect(container.querySelector('[data-testid="outcome-btn-Dave"]')).not.toBeNull();
-  });
-
-  it('clicking outcome button on non-playing player calls onDirectOutcome', () => {
-    const onDirectOutcome = vi.fn();
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} onDirectOutcome={onDirectOutcome} />
-    );
-    container.querySelector('[data-testid="outcome-btn-Bob"]').click();
-    expect(onDirectOutcome).toHaveBeenCalledWith('Bob');
-  });
-
-  it('sets background color for pending participation status', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'pending', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.style.backgroundColor).toBe('#fef08a');
-  });
-
-  it('sets background color for joined participation status', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'joined', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.style.backgroundColor).toBe('#bbf7d0');
-  });
-
-  it('sets background color for handed_back participation status', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'handed_back', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.style.backgroundColor).toBe('#fef08a');
-  });
-
-  it('formats pending status text as "pending"', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'pending', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.textContent).toContain('pending');
-  });
-
-  it('formats joined status text as "joined"', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'joined', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.textContent).toContain('joined');
-  });
-
-  it('formats handed_back status text as "handed back"', () => {
-    const players = [
-      { name: 'Zara', recorded: false, status: 'handed_back', outcomeStreet: null },
-    ];
-    const container = renderToContainer(
-      <PlayerGrid {...defaultProps} players={players} />
-    );
-    const row = container.querySelector('[data-testid="player-row-Zara"]');
-    expect(row.textContent).toContain('handed back');
-  });
-
-  it('preserves existing status colors when new statuses are added', () => {
-    const container = renderToContainer(<PlayerGrid {...defaultProps} />);
-    const rows = container.querySelectorAll('[data-testid^="player-row-"]');
-    expect(rows[0].style.backgroundColor).toBe('#ffffff'); // playing
-    expect(rows[1].style.backgroundColor).toBe('#bbf7d0'); // won
-    expect(rows[2].style.backgroundColor).toBe('#fecaca'); // folded
-    expect(rows[3].style.backgroundColor).toBe('#fed7aa'); // lost
-    expect(rows[4].style.backgroundColor).toBe('#e5e7eb'); // not_playing
-  });
-
-  describe('participation mode — sit-out button', () => {
-    const participationPlayers = [
-      { name: 'Alice', recorded: false, status: 'playing', outcomeStreet: null },
-      { name: 'Dan', recorded: false, status: 'idle', outcomeStreet: null },
-      { name: 'Bob', recorded: false, status: 'pending', outcomeStreet: null },
-      { name: 'Carol', recorded: false, status: 'handed_back', outcomeStreet: null },
-    ];
-
-    it('shows sit-out button for playing-status players in participation mode', () => {
-      const onMarkNotPlaying = vi.fn();
-      const container = renderToContainer(
-        <PlayerGrid
-          {...defaultProps}
-          players={participationPlayers}
-          gameMode="participation"
-          onMarkNotPlaying={onMarkNotPlaying}
-        />,
-      );
-      expect(container.querySelector('[data-testid="sitout-btn-Alice"]')).not.toBeNull();
-      expect(container.querySelector('[data-testid="sitout-btn-Dan"]')).not.toBeNull();
-    });
-
-    it('does not show sit-out button for non-playing-status players', () => {
-      const onMarkNotPlaying = vi.fn();
-      const container = renderToContainer(
-        <PlayerGrid
-          {...defaultProps}
-          players={participationPlayers}
-          gameMode="participation"
-          onMarkNotPlaying={onMarkNotPlaying}
-        />,
-      );
-      expect(container.querySelector('[data-testid="sitout-btn-Bob"]')).toBeNull();
-      expect(container.querySelector('[data-testid="sitout-btn-Carol"]')).toBeNull();
-    });
-
-    it('does not show sit-out button in dealer_centric mode', () => {
-      const onMarkNotPlaying = vi.fn();
-      const container = renderToContainer(
-        <PlayerGrid
-          {...defaultProps}
-          players={participationPlayers}
-          gameMode="dealer_centric"
-          onMarkNotPlaying={onMarkNotPlaying}
-        />,
-      );
-      expect(container.querySelector('[data-testid="sitout-btn-Alice"]')).toBeNull();
-    });
-
-    it('calls onMarkNotPlaying with player name when sit-out button clicked', () => {
-      const onMarkNotPlaying = vi.fn();
-      const container = renderToContainer(
-        <PlayerGrid
-          {...defaultProps}
-          players={participationPlayers}
-          gameMode="participation"
-          onMarkNotPlaying={onMarkNotPlaying}
-        />,
-      );
-      container.querySelector('[data-testid="sitout-btn-Alice"]').click();
-      expect(onMarkNotPlaying).toHaveBeenCalledWith('Alice');
-    });
-
-    it('does not show sit-out button when onMarkNotPlaying is not provided', () => {
-      const container = renderToContainer(
-        <PlayerGrid
-          {...defaultProps}
-          players={participationPlayers}
-          gameMode="participation"
-        />,
-      );
-      expect(container.querySelector('[data-testid="sitout-btn-Alice"]')).toBeNull();
-    });
-  });
-});
